+++ frontend/src/pages/Forms/CompAspects/Noises/NoiseForm/index.js
@@ -40,15 +40,17 @@ export default function NoiseForm({
   }
 
   async function handleDelete() {
-    setLoading(true);
+    if (window.confirm('Deseja excluir permanentemente esse item?')) {
+      setLoading(true);
 
-    try {
-      const response = await api.delete(`noise/${noise.id}`);
-      setLoading(false);
-      if (response.data.okay) onDeleteNoise(noise.id);
-    } catch (err) {
-      if (err.response) alert(err.response.data.error);
-      setLoading(false);
+      try {
+        const response = await api.delete(`noise/${noise.id}`);
+        setLoading(false);
+        if (response.data.okay) onDeleteNoise(noise.id);
+      } catch (err) {
+        if (err.response) alert(err.response.data.error);
+        setLoading(false);
+      }
     }
   }
