+++ frontend/src/pages/Forms/EnvironAspects/Effluents/SanitaryEffluents/EffluentForm/index.js
@@ -55,16 +55,18 @@ export default function EffluentForm({
   }
 
   async function handleDelete() {
-    setLoading(true);
+    if (window.confirm('Deseja excluir permanentemente esse item?')) {
+      setLoading(true);
 
-    try {
-      const response = await api.delete(`effluent/${effluent.id}`);
-      if (response.data.okay) onDeleteEffluent(effluent.id);
-    } catch (err) {
-      if (err.response) alert(err.response.data.error);
+      try {
+        const response = await api.delete(`effluent/${effluent.id}`);
+        if (response.data.okay) onDeleteEffluent(effluent.id);
+      } catch (err) {
+        if (err.response) alert(err.response.data.error);
+      }
+
+      setLoading(false);
     }
-
-    setLoading(false);
   }
 
   function handleQuantity(value) {
