+++ frontend/src/pages/Forms/EnvironAspects/WaterSupplies/WaterSupplyForm/index.js
@@ -51,15 +51,17 @@ export default function WaterSupplyForm({
   }
 
   async function handleDelete() {
-    setLoading(true);
+    if (window.confirm('Deseja excluir permanentemente esse item?')) {
+      setLoading(true);
 
-    try {
-      const response = await api.delete(`water-supply/${waterSupply.id}`);
-      setLoading(false);
-      if (response.data.okay) onDeleteWaterSupply(waterSupply.id);
-    } catch (err) {
-      if (err.response) alert(err.response.data.error);
-      setLoading(false);
+      try {
+        const response = await api.delete(`water-supply/${waterSupply.id}`);
+        setLoading(false);
+        if (response.data.okay) onDeleteWaterSupply(waterSupply.id);
+      } catch (err) {
+        if (err.response) alert(err.response.data.error);
+        setLoading(false);
+      }
     }
   }
