+++ frontend/src/pages/Forms/GeneralInfo/Representatives/RepresentativeForm/index.js
@@ -44,16 +44,20 @@ export default function RepresentativeForm({
   }
 
   async function handleDelete() {
-    setLoading(true);
+    if (window.confirm('Deseja excluir permanentemente esse item?')) {
+      setLoading(true);
 
-    try {
-      const response = await api.delete(`representative/${representative.id}`);
-      if (response.data.okay) onDeleteRepresentative(representative.id);
-    } catch (err) {
-      if (err.response) alert(err.response.data.error);
+      try {
+        const response = await api.delete(
+          `representative/${representative.id}`
+        );
+        if (response.data.okay) onDeleteRepresentative(representative.id);
+      } catch (err) {
+        if (err.response) alert(err.response.data.error);
+      }
+
+      setLoading(false);
     }
-
-    setLoading(false);
   }
 
   function handlePhoneNumber(e) {
