+++ frontend/src/pages/Forms/ProductiveProcess/Equipments/ProductiveEquipments/EquipmentForm/index.js
@@ -57,16 +57,18 @@ export default function EquipmentForm({
   }
 
   async function handleDelete() {
-    setLoading(true);
+    if (window.confirm('Deseja excluir permanentemente esse item?')) {
+      setLoading(true);
 
-    try {
-      const response = await api.delete(`equipment/${equipment.id}`);
-      if (response.data.okay) onDeleteEquipment(equipment.id);
-    } catch (err) {
-      if (err.response) alert(err.response.data.error);
+      try {
+        const response = await api.delete(`equipment/${equipment.id}`);
+        if (response.data.okay) onDeleteEquipment(equipment.id);
+      } catch (err) {
+        if (err.response) alert(err.response.data.error);
+      }
+
+      setLoading(false);
     }
-
-    setLoading(false);
   }
 
   function handleAmount(value) {
