+++ frontend/src/pages/Forms/ProductiveProcess/Supplies/SupplyForm/index.js
@@ -58,15 +58,17 @@ export default function SupplyForm({
   }
 
   async function handleDelete() {
-    setLoading(true);
+    if (window.confirm('Deseja excluir permanentemente esse item?')) {
+      setLoading(true);
 
-    try {
-      const response = await api.delete(`supply/${supply.id}`);
-      setLoading(false);
-      if (response.data.okay) onDeleteSupply(supply.id);
-    } catch (err) {
-      if (err.response) alert(err.response.data.error);
-      setLoading(false);
+      try {
+        const response = await api.delete(`supply/${supply.id}`);
+        setLoading(false);
+        if (response.data.okay) onDeleteSupply(supply.id);
+      } catch (err) {
+        if (err.response) alert(err.response.data.error);
+        setLoading(false);
+      }
     }
   }
